Update coding style, version method, documentation
Update the coding style in a few places where my Perl style has changed
and clean up the scripts a bit.  Switch the versioning system from the
CVS $Id$ tags to a variable that will be replaced during make dist (not
yet in place).  Proofread the documentation and update it as necessary.
Refer to the new afs-monitor distribution, which these scripts will now
be part of.