Update binary search logic to look in PATH as a fallback
The vos binary is installed in sbin by the Red Hat packages and the
default upstream installation rules, so check in sbin as well as bin.
Patch from Daniel Scott.  Also fall back, for all AFS binaries, to
using PATH if the binary can't be found in known locations.

Also checks for vos in /usr/local/sbin/vos.